refactor: rename action and resource store
