+++ src/page/App/components/ActionEditor/ActionEditorPanel/TitleInput/interface.ts
@@ -1,5 +1,5 @@
 import { HTMLAttributes } from "react"
-import { ActionItem } from "@/redux/currentApp/action/actionList/actionListState"
+import { ActionItem } from "@/redux/currentApp/action/actionState"
 
 export interface TitleInputProps
   extends Omit<HTMLAttributes<HTMLDataElement>, "title"> {
