+++ src/page/App/components/ActionEditor/Resource/RESTAPI/param/index.tsx
@@ -6,8 +6,8 @@ import { Input } from "@illa-design/input"
 import { FieldArray } from "@/page/App/components/ActionEditor/ActionEditorPanel/ResourceEditor/FieldArray"
 import { ActionEditorContext } from "@/page/App/components/ActionEditor/context"
 import { useSelector } from "react-redux"
-import { selectAllActionItem } from "@/redux/currentApp/action/actionList/actionListSelector"
-import { selectAllResource } from "@/redux/currentApp/action/resource/resourceSelector"
+import { selectAllActionItem } from "@/redux/currentApp/action/actionSelector"
+import { selectAllResource } from "@/redux/currentApp/resource/resourceSelector"
 import {
   configContainerStyle,
   descriptionStyle,
