--- src/redux/currentApp/action/actionList/actionListSelector.ts
@@ -1,4 +1,0 @@
-import { RootState } from "@/store"
-
-export const selectAllActionItem = (state: RootState) =>
-  state.currentApp.action.actionList
