+++ src/redux/currentApp/action/actionSelector.ts
@@ -1,0 +1,3 @@
+import { RootState } from "@/store"
+
+export const selectAllActionItem = (state: RootState) => state.currentApp.action
