+++ src/redux/currentApp/resource/resourceReducer.ts
@@ -2,7 +2,7 @@ import { CaseReducer, PayloadAction } from "@reduxjs/toolkit"
 import {
   ResourceListState,
   Resource,
-} from "@/redux/currentApp/action/resource/resourceState"
+} from "@/redux/currentApp/resource/resourceState"
 
 export const addResourceItemReducer: CaseReducer<
   ResourceListState,
