+++ src/redux/currentApp/resource/resourceSelector.ts
